+++ semana18/template-labook/src/model/Post.ts
@@ -1,0 +1,13 @@
+export enum POST_TYPES {
+    NORMAL = "normal",
+    EVENT = "event"
+}
+ 
+export type Post = {
+    id: string,
+    photo: string,
+    description: string,
+    type: POST_TYPES,
+    createdAt: Date,
+    authorId: string
+}
