+++ semana18/template-labook/src/model/User.ts
@@ -1,0 +1,12 @@
+ export type User = {
+    id: string,
+    name: string,
+    email: string,
+    password: string,
+    role: USER_ROLES
+}
+
+export enum USER_ROLES {
+    NORMAL = 'NORMAL',
+    ADMIN = 'ADMIN'
+}
