+++ semana18/template-labook/src/routes/userRoutes.ts
@@ -1,0 +1,8 @@
+import express from "express";
+
+export const userRouter = express.Router();
+
+// userRouter.post("/signup", createUser);
+// userRouter.post("/login", login);
+// userRouter.post("/edit", editUser);
+// userRouter.get("/:id", getUserById);
