+++ semana18/template-labook/src/services/idGenerator.ts
@@ -1,0 +1,5 @@
+import { v4 } from "uuid"
+
+export const generateId = (): string => {
+   return v4()
+}
